add: additional info last update add: button disable when additional info modal is open add: button loading when additional info is fetching
